Fix analog input constants;
analogRead takes in the number of the analog pin which is converted
to an ADC channel with analogPinToChannel. The current mapping is a
direct 1 to 1 mapping where the analog pin number is the same as the
ADC channel. The values were previously set as the pin numbers on the
microcontroller and would thus not work.